% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Dataset}
\format{
A data frame with 144 rows and 6 columns:
\describe{
  \item{RIVER}{River where fish was observed}
  \item{MILE}{Mile marker wehre fish was observed}
  \item{SPECIES}{Species of observed fish}
  \item{LENGTH}{Length of fish}
  \item{WEIGHT}{Weight of fish}
  \item{DDT}{concentration of DDT in fish}
}
}
\source{
<https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
ddt
}
\description{
A dataset containing information about DDT concentrations in fish species across different rivers.
}
\keyword{datasets}
