% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstrap Confidence Interval and Histogram}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations. Default is 10000.}

\item{x}{A numeric vector of data for bootstrap resampling.}

\item{fun}{A function to compute the statistic of interest. The default is mean.}

\item{alpha}{Significance level for confidence intervals. Default is 0.05.}

\item{cx}{Scaling factor for text size in the plot. Default is 1.5.}

\item{...}{Additional graphical parameters passed to the 'hist()' function.}
}
\value{
A list containing:
\item{ci}{Confidence interval for the statistic.}
\item{fun}{The function used for computation.}
\item{x}{The original data vector.}
}
\description{
This function performs bootstrap resampling to compute confidence intervals and visualize the distribution of the sample statistics.
}
\examples{
\dontrun{
set.seed(39); sam=rnorm(25,mean=25,sd=10)
myboot(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)
}
}
