% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Find Maximum Likelihood Estimate}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A function representing the log-likelihood. It should accept two arguments: data points('x') and parameter values ('param').}

\item{x}{A numeric vector of data points.}

\item{param}{A numeric vector of parameter values to evaluate the log-likelihood function.}

\item{...}{Additional graphical parameters for the plot.}
}
\value{
A list containing:
\item{i}{Index of the parameter value that maximizes the likelihood.}
\item{parami}{The parameter value corresponding to the maximum likelihood.}
\item{yi}{The maximum log-likelihood value.}
\item{slope}{The slopes around the maximum likelihood point, if applicable.}
}
\description{
This function calculates the maximum likelihood estimate (MLE) for a given log-likelihood function, a set of data points, and a range of parameter values. It visualizes the likelihood function and identifies the parameter value that maximizes it.
}
\examples{
\dontrun{
# Example usage with a binomial log-likelihood function
logbin <- function(x, param) dbinom(x, size = 10, prob = param, log = TRUE)
v <- mymaxlik(
  x = c(9, 10, 11),
  param = seq(0, 1, length = 1000),
  lfun = logbin,
  xlab = expression(p),
  main = "Binomial MLE for p",
  cex.main = 2
)
}
}
