% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Log-Likelihood Function for Three Binomial Experiments}
\usage{
mymaxlikg(lfun = "logbin3", theta)
}
\arguments{
\item{lfun}{A character string specifying the name of the log-likelihood function to be used.  The default is "logbin3".}

\item{theta}{A numeric vector of probability values to evaluate the likelihood function.}
}
\value{
The value of theta corresponding to the maximum likelihood.
}
\description{
Computes the log-likelihood for three binomial experiments given a probability parameter theta.
}
\examples{
\dontrun{
logbin3 <- function(theta){
log(dbinom(4, prob = theta, size = 10)) +  # X1 = 4/10
log(dbinom(12, prob = theta, size = 20)) + # X2 = 12/20
log(dbinom(7, prob = theta, size = 15))    # X3 = 7/15
}

mymaxlikg(theta=seq(0,1,length=10000))
}
}
