% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlnorm.R
\name{mymlnorm}
\alias{mymlnorm}
\title{Maximum Likelihood Estimation for Normal Distribution Parameters}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{Numeric vector. The sample data for which the parameters are estimated.}

\item{mu}{Numeric vector. A range of candidate values for the mean.}

\item{sig}{Numeric vector. A range of candidate values for the standard deviation.}

\item{...}{Additional graphical parameters to be passed to the contour function.}
}
\value{
A list containing:
\item{x}{The input sample data.}
\item{muest}{The estimated mean based on MLE.}
\item{sigest}{The estimated standard deviation based on MLE.}
\item{maxl}{The maximum likelihood value.}
}
\description{
This function estimates the mean and standard deviation of a normal distribution using maximum likelihood estimation based on a given sample. It also provides a contour plot of the likelihood function.
}
\examples{
\dontrun{
x <- c(5, 7, 7, 8, 10)  # Sample data
mu <- seq(4, 12, length.out = 100)  # Grid of μ values
sig <- seq(0.1, 5, length.out = 100)  # Grid of σ values

mymlnorm(x, mu, sig)
}
}
