% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynewt.R
\name{mynewt}
\alias{mynewt}
\title{Newton-Raphson Method for Root Finding}
\usage{
mynewt(x0, delta = 0.001, f, h = 0.001, xlim = c(-x0, x0))
}
\arguments{
\item{x0}{Initial guess for the root.}

\item{delta}{Convergence threshold; the iteration stops when the absolute value of the function at the current approximation is less thanhis value. Default is 0.001.}

\item{f}{A function whose root is to be found.}

\item{h}{Step size for numerical differentiation. Default is 0.001.}

\item{xlim}{Range of x-values for plotting. Default is c(-x0, x0).}
}
\value{
A list containing:
\item{x}{Vector of approximations for the root at each iteration.}
\item{y}{Vector of function values corresponding to each approximation.}
}
\description{
Implements the Newton-Raphson algorithm to find the root of a given function. The method iteratively approximates the root by using the derivative of the function.
}
\examples{
\dontrun{
# Find the root of a quadratic function
mynewt(x0 = 10, delta = 0.000001, f = function(x) x^2 - 5*x + 6)
}
}
