% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson Method for Log-Likelihood Optimization}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{Initial guess for the root.}

\item{delta}{Step size for numerical differentiation. Default is 0.001.}

\item{llik}{Log-likelihood function to be optimized.}

\item{xrange}{Range of values for plotting the log-likelihood and its derivative.}

\item{parameter}{Name of the parameter being optimized (used for axis labels in plots). Default is "param".}
}
\value{
A list containing:
\item{x}{Vector of successive approximations to the root.}
\item{y}{Vector of derivative values at each iteration.}
}
\description{
This function implements the Newton-Raphson method to optimize the log-likelihood function. It iteratively find better approximations to the roots of the derivative of the log-likelihood function.
}
\examples{
\dontrun{
# Example using Poisson log-likelihood:
myNRML(x0 = 2, delta = 0.000001,
       llik = function(x)
       log(dpois(3, lambda = x) * dpois(4, lambda = x)),
       xrange = c(0.01, 10),
       parameter = "lambda")
}
}
